Add topic_id returned by storage to the TopicMetadata
- Rationale.
 To be able to distinguish topics by unique topic ID rather than by
 topic name in the future.

+++ rosbag2_storage_sqlite3/include/rosbag2_storage_sqlite3/sqlite_wrapper.hpp
@@ -50,7 +50,7 @@ public:
   SqliteStatement prepare_statement(const std::string & query);
   std::string query_pragma_value(const std::string & key);
 
-  size_t get_last_insert_id();
+  int64_t get_last_insert_id();
 
   operator bool();
